+++ pkg/templates/parser_stats.go
@@ -8,5 +8,6 @@ const (
 	HeadlessFlagWarningStats = "headless-flag-missing-warnings"
 	TemplatesExecutedStats   = "templates-executed"
 	CodeFlagWarningStats     = "code-flag-missing-warnings"
+	FuzzFlagWarningStats     = "fuzz-flag-missing-warnings"
 	SkippedUnsignedStats     = "skipped-unsigned-stats" // tracks loading of unsigned templates
 )
